% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_clust.R
\name{dimred_clust}
\alias{dimred_clust}
\title{dimred_clust}
\usage{
dimred_clust()
}
\arguments{
\item{x}{input value}
}
\value{
k clusters
}
\description{
Performs dimensionality reduction and clustering
}
\examples{


}
