% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heir_clust.R
\name{heir_clust}
\alias{heir_clust}
\title{heir_clust}
\usage{
heir_clust()
}
\arguments{
\item{x}{input value}
}
\value{
k clusters
}
\description{
Clusters data using heirarchical clustering algorithm
}
\examples{


}
