% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clust.R
\name{kmeans_clust}
\alias{kmeans_clust}
\title{kmeans_clust}
\usage{
kmeans_clust()
}
\arguments{
\item{x}{input value}
}
\value{
k clusters
}
\description{
Clusters data according to k-means algorithm
}
\examples{


}
