% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMixEm.R
\name{normalmix_init}
\alias{normalmix_init}
\title{Initialization for normMixEm Algorithm}
\usage{
normalmix_init(x, k = 5)
}
\arguments{
\item{x}{A matrix of size n*p consisting of the data.}

\item{k}{Number of components. Default as 5.}
}
\value{
A list of parameters for initialization including:
\itemize{
  \item lambda - A vector of length k includes the probability parameters of k components.
  \item mu - A k*p matrix includes the mean parameters of k Normals.
  \item sigma - A k*p matrix includes the variance parameters of k Normals.
  \item k - Integer. Number of components.  
}
}
\description{
Internal initialization function for normMixEm algorithm in this package. Modified based on the normalmix.init{mixtools}.
}
