% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMixEm.R
\docType{class}
\name{normMixEM-class}
\alias{normMixEM-class}
\alias{normMixEm}
\title{normMixEm}
\arguments{
\item{input_dat}{n x p data matrix}

\item{num_components}{number of clusters}
}
\value{
A list  
\itemize{
  \item convergence
  \item mu_mat
  \item sigma_mat
  \item pi_vec
  \item iter
  \item loglik+list
  \item prob_mat
}
}
\description{
?Description
}

