% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMixEm.R
\docType{class}
\name{normMixEM}
\alias{normMixEM}
\title{EM Algorithm for Mixtures of Independent Multivariate Normals}
\arguments{
\item{input_data}{A matrix/dataframe of size nxp consisting of the data.}

\item{num_components}{Number of components.}
}
\value{
A list including
\itemize{
  \item convergence - Convergece status. 0 if converged,1 if not.
  \item mu_mat - A k*p matrix includes the mean parameters of k normals.
  \item sigma_mat - A k*p matrix includes the variance parameters of k normals.
  \item pi_vec - A vector of length k includes the probability parameters of k components.
  \item iter - A integer. Number of iterations
  \item loglik_list - A number. The log-likelihood of the output result
  \item prob_mat - A n*k matrix includes the probabilities of being assigned to k components for each sample.
}
}
\description{
This is a normMixEM class, which initializes and runs EM algorithm output for mixtures of independent multivariate normal distributions
}

