% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMixEm.R
\name{rmvnorm_chol}
\alias{rmvnorm_chol}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
rmvnorm_chol(n, mu = NULL, sigma = NULL)
}
\arguments{
\item{n}{Number of vectors to simulate.}

\item{mu}{Mean vector.}

\item{sigma}{Covariance matrix, assumed symmetric and nonnegative definite.}
}
\value{
A n*p matrix. Each row is a multivariate normal vector generated independently based on the input parameters.
}
\description{
Sample from a multivariate normal distribution based on Cholesky decomposition. Modified based on rmvnorm {mixtools}
}
