% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC}
\alias{TC}
\title{scRNA Gene Expression Example Labels}
\format{
A 801-length vector
\describe{
  \item{label}{one of BRCA, KIRC, COAD, LUAD, PRAD}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/gene+expression+cancer+RNA-Seq#}
}
\usage{
TC
}
\description{
Labels for cancer types that correspond to the rows of tumor_reduced.
}
\keyword{datasets}
