% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_reduced}
\alias{tumor_reduced}
\title{scRNA Gene Expression Example Data}
\format{
A data frame with 801 rows and 2000 variables:
\describe{
  \item{gene_xxxx}{gene expression data for gene_xxxx for each sample}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/gene+expression+cancer+RNA-Seq#}
}
\usage{
tumor_reduced
}
\description{
scRNA data with labels for cancer types (from the UCI ML repository).
    Each row is a sample, and each column is a gene.
}
\keyword{datasets}
